LF-9762: drm: bridge: mhdp hdmi: Fix ELD data is not update issue
commit  3124f7e445a37e1897c1daf72e5fc430d63ed373 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The ELD (EDID-Like Data) is not updated when the HDMI cable
is plugged into different HDMI monitors.
This is because the EDID is not updated in the HDMI HPD function.
As a result, the ELD data remains unchanged and may not reflect
the capabilities of the newly connected HDMI sink device.

To address this issue, the handle_plugged_change function should move to
the bridge_atomic_enable and bridge_atomic_disable functions.
Make sure the EDID is properly updated before updating ELD

Signed-off-by: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>