i2c: designware: introduce a custom scl recovery for SoCFPGA platforms
commit 29ba509f64390caaa357f2c99aebf1c5063e2068 from
https://github.com/altera-fpga/linux-socfpga socfpga-6.12.11-lts.

The I2C pins on the SoCFPGA platforms do not go through a GPIO module,
thus cannot be recovered by the default method of by doing a GPIO access.
Only a reset of the I2C IP block can a recovery be successful, so this
change effectively resets the I2C controller, NOT any attached clients.

Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Dinh Nguyen <dinguyen@kernel.org>
Signed-off-by: Liwei Song <liwei.song@windriver.com>