drivers/nvmem: s32cc-gpr-nvmem: Separate platform-specific offset handlers
commit d1831a983e170dea31bfdee159a571704332a7b5 from
https://github.com/nxp-auto-linux/linux

To avoid the situation where a platform uses an NVMEM read/write
handler with an offset for a cell not corresponding to itself,
but to another platform, separated the handlers for each category
of cells (per platform), which are now called from the most
specific one to the most generic one (e.g.: s32r45 calls s32cc,
if an s32r45-specific cell offset is not matched).

The first handler to be called is determined by the compatible
against which the driver is matched when probed.

Issue: ALB-10366
Issue: ALB-10893
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>