dt-bindings: interrupt-controller: document PolarFire SoC's gpio interrupt mux
commit 5f29c43eebe3959cb8247d9cd59228054dc32726 from
https://github.com/linux4microchip/linux.git

On PolarFire SoC there are more GPIO interrupts than there are interrupt
lines available on the PLIC, and a runtime configurable mux is used to
decide which interrupts are assigned direct connections to the PLIC &
which are relegated to sharing a line.
This mux is, in our reference configuration, written by platform
firmware during boot based on the FPGA's configuration.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>