arm64: dts: ls1046a: add SerDes nodes
commit  59000f0e58708fa2db0d462efdcacd325b6b85d1 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Add the OF nodes for the 2 SerDes blocks on this SoC, and tentatively
also add the DCFG_DCSR region which will be necessary in the future for
RCW override (for dynamic SerDes protocol change). That is not currently
implemented.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>