spi: spi-fsl-lpspi: Add "cs_change" support
commit 5997b186c7988401c1e9edaf15962d45c2a0d39b from
https://github.com/nxp-auto-linux/linux

According to the documentation, SR.FCF is set when frame is complete, and
PCS has negated. If cs_change is set and the transfer is the last one in
the message, PCS will remain asserted. In this case, SR.WCF will be set
instead of SR.FCF. The interrupt logic was changed to handle this case
also.

Issue: ALB-10628
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
[Zhantao: use spi_get_chipselect(spi, 0) to replace the original
spi->chip_select]
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>