net: stmmac: Remove redundant checking for rx_coalesce_usecs
commit b4ab96b8701749e6f25b915ecb99f90f4e2e5fe7 from
https://github.com/intel/linux-intel-lts.git

The datatype of rx_coalesce_usecs is u32, always larger or equal to zero.
Previous checking does not include value 0, this patch removes the
checking to handle the value 0. This change in behaviour making the
value of 0 cause an error is not a problem because 0 is out of
range of rx_coalesce_usecs.

Signed-off-by: Gan Yi Fang <yi.fang.gan@intel.com>
Link: https://lore.kernel.org/r/20231018030802.741923-1-yi.fang.gan@intel.com
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>