crypto: jitter - use permanent health test storage
commit 1d1d3c0f8471c3825425c144f4f4df4688672a68 from
git@git.assembla.com:cavium/WindRiver.linux.git

The health test result in the current code is only given for the currently
processed raw time stamp. This implies to react on the health test error,
the result must be checked after each raw time stamp being processed. To
avoid this constant checking requirement, any health test error is recorded
and stored to be analyzed at a later time, if needed.

This change ensures that the power-up test catches any health test error.
Without that patch, the power-up health test result is not enforced.

The introduced changes are already in use with the user space version of
the Jitter RNG.

Fixes: 04597c8dd6c4 ("crypto: jitter - add RCT/APT support for different OSRs")
Reported-by: Joachim Vandersmissen <git@jvdsn.com>
Closes: https://jvdsn.com/blog/linux-kernel-jitterentropy-rng-weak-keys/
Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Prince John <pjohn@marvell.com>
Change-Id: I4b4b8b06f78996ef1e3ad048e330ae10b3d5cf9f
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/157561
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>