mailbox: llce: Allocate fifos in ascending order
commit 48a29a4c4924939de421549184567ea1d5d3f533 from
https://github.com/nxp-auto-linux/linux

This commit tries to mitigate the impact of the LLCE FW issue in which
the TX ACK frames are sent to the wrong LLCE CAN interface. This
scenario is reproduced when the fifo id allocated for a controller is
not the same with the controller id.
Allocating the fifo ids in ascending error tries to limitate the impact
of the firmware issue.

Issue: ALB-10677
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>