arm_scmi: smc: save flags only if spin lock was acquired
commit c298e8a0e1d63dff0146331a5d50f98bc0f099d6 from
https://github.com/nxp-auto-linux/linux

Previously, if we had concurrent requests, we could end up in a
situation were incorrect flags were restored which resulted in a
core running with interrupts disabled. Then, schedutil would try to
run and discover that interrupts were disabled and it would, in turn,
issue the bug message.

Issue: ALB-10746
Issue: ALB-10894

Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>