gpio: use pinctrl_gpio_get_mux_owner
commit 412ce27018cda600fdcb479ebe495dd9bc42104a from
https://github.com/nxp-auto-linux/linux

The `desc->label` is only filled with the name of the GPIO owner.
Therefore, commands such as `gpioinfo` would return that a pin is free
but that doesn't mean the pin is free, it just means it isn't requested
as a GPIO. By calling the before mentioned function, we will also
retrieve if a mux_owner exists (assigned when configuring pinctrl at
boot).

Issue: ALB-10523
Issue: ALB-10876
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>