drivers: firmware: xilinx: Add sysfs to get last reset reason
commit 6aafd2def446c9fda2c4e9c0c47393ae73fd9b70 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.18

Add sysfs interface to get the last reset reason of the system from
a user space.

When the system reset happens whether intentionally or due to some
errors, the firmware used to store the last reset reason and the same
be queried by using the IOCTL API or the sysfs entry to know the
reason behind the system reset and accordingly, the error can be
diagnostic.

Signed-off-by: Ronak Jain <ronak.jain@amd.com>
Link: https://lore.kernel.org/r/20231121112056.1762641-3-ronak.jain@amd.com
State: waiting
[michals:
v6.18 update
Subject update
Rework around split
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>