ACPI / PPTT: Provide a helper to walk processor containers
commit 244fa1e8b1d143962467b7b90f87bf68f5870648 from
git@git.assembla.com:cavium/WindRiver.linux.git

The PPTT describes CPUs and caches, as well as processor containers.

To enable PPI partitions, the irqchip driver needs to know how many
partitions the platform has, and which CPUs belong to which partition.
When a percpu interrupt is registered, the partition is provided to allow
a different driver to request the same percpu interrupt intid, one per
partition.

The acpi_id of the Processor Container is the natural way to do this
with ACPI, but the DSDT AML interpreter is not available early enough
for the irqchipi driver. Fortunately, the same information can be
described in the PPTT.

Add a helper to count the number or Processor Containers in the PPTT.
This is structured as a walker/callback as the irqchip driver will
also use this to configure each partition.

Only Processor entries in the PPTT that have a valid acpi id are considered
as containers. To identify a particular Processor Container, it must have
an id.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>