fpga: Add xilinx AFI config driver
commit dbe5fae0c12ebd100570072982d438473871a5e6 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Add zynqmp AXI FIFO interface(AFI) config driver. This is useful for the
configuration of the PS-PL interface on Zynq US+ MPSoC platform.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Link: https://lore.kernel.org/linux-arm-kernel/20220824035542.706433-1-nava.kishore.manne@amd.com/
State: not-upstreamable
Reason: The existing AFI interface handling is done using an FPGA subsystem
but the maintainer suggested to use an interconnect subsystem.
[michals:
v6.12 update
fpga: Fix the reset handling
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>