bindings: fpga: Add binding doc for the Zynq AFI interface
commit c163ea2242793cbac0f5fffd116e14530e414e27 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Xilinx Zynq platform connect the PS to the programmable logic(PL)
through the AXI port. This AXI port helps to establish the data path
between the PS and PL. In-order to establish the proper communication
data path between PS and PL the AXI port data path should be configured
with the proper Bus-width values.

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: not-upstreamable
Reason: The existing AFI interface handling is done using an FPGA subsystem
but the maintainer suggested to use an interconnect subsystem.
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>