fpga: zynq: Add AFI config driver
commit ac1a40ce00849a567cf3248fdb6991bbf0d237a5 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Add Zynq AXI FIFO interface(AFI) config driver. This is useful for the
configuration of the PS-PL interface on Zynq platform.

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>
Tested-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: not-upstreamable
Reason: The existing AFI interface handling is done using an FPGA subsystem
but the maintainer suggested to use an interconnect subsystem.
[michals: v6.12 update
fpga: zynq-afi: Resolve kernel-doc warning in AFI struct description
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>