drivers: i2c: octeon: Avoid potential infinite loop
commit 716c204c485d7df58e7b9198ea60062a99f1825e from
git@git.assembla.com:cavium/WindRiver.linux.git

The time_left variable is used and set to 1000, which is not
being checked in while loop condition - this could trigger as
infinite loop in case the test flag is not set. Adding time_left
check in while condition as a way to exit the loop in that case.

Signed-off-by: Piyush Malgujar <pmalgujar@marvell.com>
Change-Id: I9dad8ace04dd1ebb06ae19b1f4ec546d71042b32
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/117996
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>