usb: dwc3: gadget: ISOC transfers should be stopped before starting a transfer
commit ddd50ba51153b1fae42962847c91e8e12ea8e890 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

For ISOC transfers the requests are not queued until the HOST requests
for data and XferNotReady event is generated .But XferNotReady event
is not getting generated for ISOC transfers for the second time after
Endpoint configuration. Since ISOC packets depend on the XferNotReady
events, they will not be queued to controller. Because of this issue
timeout happens on the application layer.

This patch fixes this issue by issuing END TRANSFER command before
starting any ISOC transfers. Doing so will make the controller clear
the previous allocated endpoint resources and reallocate resources
when the transfer is requested. Because of this change XferNotReady
events will be generated when host requests for the ISOC transfer.

Signed-off-by: Manish Narani <manish.narani@xilinx.com>
Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>