dpaa2-switch: change dpaa2_switch_port_set_fdb() function prototype
commit  f8bf965a7499d8efafd83058b5dbb6955e20d9c8 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Since there dpaa2_switch_port_set_fdb() never fails and its return value
was never checked, change its prototype to return void.

Also, instead of determining if the DPAA2 port is joining or leaving an
upper based on the value of the 'bridge_dev' parameter, add the
'linking' parameter to explicitly specify the action. This will enable
us to pass the upper device that we are joining/leaving in all possible
cases. This will get used in the next patches to determine what kind of
device the upper is: a bridge or a bond.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>