net: tsn: remove unused notifier chain functionality
commit  05fdcb2a21f53e14446222503bae70a368b9b12e from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Apparently it was intended for drivers to emit TSN_QBV_CONFIGCHANGETIME_ARRIVE
and for this to notify user space, but it seems to have never been
implemented.

If it ever is, it is rather likely we will want to expose this through
tc instead, so just drop the unused code from the legacy tsntool.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>