soc: fsl: guts: make it easier to determine on which SoC we are running
commit  91c217a405f71a1e4840de081253a19df3c57597 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Add a new qoriq_die enum to be used as an index in the fsl_soc_die
array. A new fsl_soc_die_match_one() function is also added so that we
can directly determine if the SVR is a match with a specific DIE.
The SVR value read from the DCFG CCSR is also kept in the global soc
structure so that it can be accessed when needed.

These changes are useful because the guts driver will need to easily
determine on which SoC it's running in the next patches.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>